+++ R/data-facelayer-resample.R
@@ -15,16 +15,16 @@
 #' # create a grid
 #' g <- trigrid(c(4,4))
 #' # create a data layer
-#' fl <- facelayer(g)
-#' fl@values<-rnorm(length(fl))
+#' fl <- facelayer(g, rnorm(length(faces(g))))
 #' # target structure
 #' h <- trigrid(4)
 #' # resampling
 #' res <- resample(fl, h)
-#' fl2<-facelayer(h)
+#' fl2<-facelayer(h, res)
 #' fl2@values[] <- res
 #'
 #' @exportMethod resample
+#'
 #' @rdname resample
 setMethod(
 	"resample",
@@ -162,13 +162,13 @@ setMethod(
 	definition=function(x, y,method=NULL,res=5){
 		# make a copy of the original entity
 		newRast <- y
-		
+
 		# get raster coordinates
-		allCoords <- as.data.frame(terra::xyFromCell(y, 1:terra::ncell(y)))
+		allCoords <- as.data.frame(terra::xyFromCell(newRast, 1:terra::ncell(newRast)))
 
 		# make the sf with crs
 		sfPoints <- sf::st_as_sf(
-			as.data.frame(allCoords), coords=c("x", "y"), crs=terra::crs(y))
+			as.data.frame(allCoords), coords=c("x", "y"), crs=terra::crs(newRast))
 		# locate the points
 		tiedGrid <- dynGet(x@grid, minframe=0L)
 
@@ -176,250 +176,26 @@ setMethod(
 		cells <- locate(tiedGrid, sfPoints)	
 
 		# get the appropriate values
-		terra::values(newRast) <- values(x)[cells]
+		vals <-x@values
+		names(vals) <- x@names
+		terra::values(newRast) <- vals[cells]
 
 		# return the thing
 		return(newRast)
 
 })
 
+#' @rdname
+setMethod(
+	"resample",
+	signature=c("SpatRaster", "facelayer"),
+	definition=function(x,y,...){
+		# get the grid of the facelayer
+		tiedGrid <- dynGet(y@grid, minframe=0L)
+		# do a resample
+		resample(x, y=tiedGrid, ...)
 
-#' Icosahedral grid-based density estimation
-#'
-#' Spatial density estimation algorithm based on rotation of icosahedral grids.
-#'
-#' Any points set can be binned to an icosahedral grid (i.e. number of incidences can be counted), which will be dependent on the exact positions of grid cells. Rotating the grid in 3d space will result in a different distribution of counts. This distribution can be resampled to a standard orientation structure. The size of the icosahedral grid cells act as a bandwidth parameter.
-#'
-#' The implemented algorithm 1) takes a point cloud (\code{x})) and an icosahedral grid \code{y} 2) randomly rotates the icosahedral grid, 3) looks up the points falling on grid cells, 4) resamples the grid to a constant orientation object (either \code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link[terra:rast]{SpatRaster}}). Steps 2-4 are repeated \code{trial} times, and then \code{FUN} is applied to every vector of values that have same spatial position.
-#' 
-#' @param x Matrix of longitude, latitude data, \code{\link[sf:sf]{sf}} class, or \code{\link[sp:SpatialPoints]{SpatialPoints}} Point cloud.
-#' @param y \code{\link{trigrid}} or \code{\link{hexagrid}} An icosahedral grid.
-#' @param out \code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link[terra:rast]{SpatRaster}}output structure.
-#' @param trials \code{numeric} value, the number of iterations.  
-#' @param FUN \code{function} The function to be applied on the iteration results.
-#' @return Either named numeric vector, or a \code{\link[terra:rast]{SpatRaster}} object. If FUN is set to \code{NULL}, the output will be either a \code{matrix} or \code{\link[terra:rast]{SpatRaster}}.
-#' @examples
-#' # example to be run if terra is present
-#' if(requireNamespace("terra", quietly=TRUE)){
-#' 
-#'  # randomly generated points
-#'  x <- rpsphere(100, output="polar")
-#' 
-#'  # bandwidth grid
-#'  y <- hexagrid(deg=13)
-#' 
-#'  # output structure
-#'  out <- terra::rast(res=5)
-#'
-#'  # the function
-#'  o <- gridensity(x, y, out, trials=7)
-#' 
-#'  # visualize results
-#'  terra::plot(o)
-#'  points(x, pch=3)
-#' }
-#'
-#' @rdname grapply
-#' @export
-gridensity <- function(x, y, out, trials=100, FUN=mean){
-	# find function
-	FUN <- match.fun(mean)
-
-	if(inherits(out, "SpatRaster")){
-		if(!requireNamespace("terra", quietly=T)) stop("This module requires the 'terra' package.")
-
-		for(i in 1:trials){
-			# execute a random rotation
-			newY <- rotate(y)
-
-			# locate the points on the rotated grid 
-			cells <- locate(newY, x)
-
-			# tabulate their counts
-			cellTab <- table(cells)
-
-			# put them on a facelayer
-			fl<- facelayer(newY)
-			fl[] <- cellTab
-
-			# resample to output
-			newZ <- resample(fl, out)
-			terra::values(newZ)[is.na(terra::values(newZ))] <- 0
-
-			if(i == 1){
-				stack <- newZ
-			}else{
-				stack <- c(stack, newZ)
-			}
-			cat(i, "\r")
-			flush.console()
-		}
-
-		if(!is.null(FUN)){
-			res <- terra::app(stack, fun=FUN)
-		}else{
-			res <- stack
-		}
 	}
 
-	if(inherits(out, "trigrid")){
-		stack <- matrix(NA, ncol=nrow(out@faces), nrow=trials)
-		colnames(stack)  <- rownames(out@faces)
-
-		for(i in 1:trials){
-			# execute a random rotation
-			newY <- rotate(y)
-
-			cells <- locate(newY, x)
-			cellTab <- table(cells)
-			fl<- facelayer(newY)
-			fl[] <- cellTab
-
-			newZ <- resample(fl, out)
-			newZ[is.na(newZ)] <- 0
-			
-			stack[i,names(newZ)] <- newZ
-			cat(i, "\r")
-			flush.console()
-		}
-
-		# calculate the thing for all of them
-		if(!is.null(FUN)){
-			res <- apply(stack, 2, FUN=FUN)
-		}else{
-			res <- stack
-		}
-	}
-	return(res)
-
-}
-
-
-#' Grid Rotation Apply
-#'
-#' Discretization with iteratively rotated icosahedral grids
-#'
-#' Simple discretization of spatial data can be subject to error due to the random assignment to grid cells. \code{grapply} offers a framework for the Monte Carlo estimation of the expectation of function that normally can be applied to
-#' a discretized set of points
-#' @param x Matrix of longitude, latitude data, \code{\link[sf:sf]{sf}} class, or \code{\link[sp:SpatialPoints]{SpatialPoints}} Point cloud.
-#' @param y \code{\link{trigrid}} or \code{\link{hexagrid}} An icosahedral grid.
-#' @param out \code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link[terra:rast]{SpatRaster}}output structure.
-#' @param trials \code{numeric} value, the number of iterations.
-#' @param FUN \code{function} The function to be applied on the iteration results. It defaults to the number of points, but it can be used to
-#' @param APP \code{function} The function to be applied on the iteration results. If set to \code{NULL}, it will return the stack of results for subsequent processing.
-#' @examples
-#' # example to be run if terra is present
-#' if(requireNamespace("terra", quietly=TRUE)){
-#'
-#'  # randomly generated points
-#'  x <- rpsphere(100, output="polar")
-#'
-#'  # bandwidth grid
-#'  y <- hexagrid(deg=13)
-#'
-#'  # output structure
-#'  out <- terra::rast(res=5)
-#'
-#'  # the function
-#'  o <- grapply(x, y, out, iter=7, FUN=nrow)
-#'
-#'  # visualize results
-#'  terra::plot(o)
-#'  points(x, pch=3)
-#' }
-#'
-#' @export
-#' @rdname grapply
-setGeneric(
-	name="grapply",
-	def=function(x,out, ...){
-		standardGeneric("grapply")
-	}
 )
 
-#' @rdname grapply
-setMethod(
-	"grapply",
-	signature=c(x="data.frame", out="SpatRaster"),
-	definition=function(x, out, y, coords=c("long", "lat"), iter=100, FUN=nrow, APP=mean, miss=NA, APP.args=NULL){
-
-		# function to iterate
-		ITER <- match.fun(FUN)
-
-		# the averaging function
-		if(!is.null(APP)) APP <- match.fun(APP)
-
-		# protype run-> enforce single argument return
-		if(!requireNamespace("terra", quietly=T)) stop("This function requires the 'terra' package.")
-
-		# the coordinates
-		coordDat <- x[, coords]
-
-		# iterate
-		for(i in 1:iter){
-			# execute a random rotation
-			newY <- rotate(y)
-
-			# locate the points on the rotated grid
-			cells <- locate(newY, coordDat)
-
-			# redefine object
-			thisX <- cbind(x, cell=cells)
-
-			# the function
-			current <- ITER(thisX)
-
-			# put them on a facelayer
-			fl<- facelayer(newY)
-			fl[] <-current
-
-			# resample to output
-			newZ <- resample(fl, out)
-			if(!is.na(miss)){
-				terra::values(newZ)[is.na(terra::values(newZ))] <- miss
-			}
-
-			if(i == 1){
-				stack <- newZ
-			}else{
-				stack <- c(stack, newZ)
-			}
-			cat(i, "\r")
-			flush.console()
-		}
-
-		if(!is.null(APP)){
-			# the configures
-			conf <- list(x=stack, fun=APP)
-
-			# the provided arguments
-			conf <- c(conf, APP.args)
-
-			# do a call
-			res <- do.call(terra::app, conf)
-
-		}else{
-			res <- stack
-		}
-
-		return(res)
-})
-
-#' @rdname grapply
-setMethod(
-	"grapply",
-	signature=c(x="matrix", out="SpatRaster"),
-	definition=function(x, out, y=NULL, z, iter=100, FUN=nrow, APP=mean){
-
-
-})
-
-
-
-#' @rdname grapply
-setMethod(
-	"grapply",
-	signature=c(x="matrix", out="trigrid"),
-	definition=function(x, out, y, z, iter=100, FUN=nrow, APP=mean){
-
-})